requirements.txt
streamlit
python-dotenv
google-generativeai
pinecone
beautifulsoup4
requests